+++ NEWS.md
@@ -1,7 +1,7 @@
 # santoku (development version)
 
-* New `raw` parameter to `chop()` overrides `raw` in `lbl_*` functions.
-  - In future, `raw` in `lbl_*` functions may be deprecated.
+* New `raw` parameter to `chop()`.
+  - `raw` in `lbl_*` functions is soft-deprecated.
 
 # santoku 0.8.0